chore: move shape and style controls in `./v1` and `./v2` directories
