+++ packages/controls/src/controls/style/v1/style.ts
@@ -1,21 +1,21 @@
 import { match, P } from 'ts-pattern'
 import { z } from 'zod'
 
-import { mapValues } from '../../lib/functional'
-import { safeParse, type ParseResult } from '../../lib/zod'
+import { mapValues } from '../../../lib/functional'
+import { safeParse, type ParseResult } from '../../../lib/zod'
 
-import { Schema, type ResponsiveValue } from '../../common'
-import { responsiveValue } from '../../common/schema'
-import { type CopyContext } from '../../context'
-import { Targets, type IntrospectionTarget } from '../../introspection'
+import { Schema, type ResponsiveValue } from '../../../common'
+import { responsiveValue } from '../../../common/schema'
+import { type CopyContext } from '../../../context'
+import { Targets, type IntrospectionTarget } from '../../../introspection'
 import {
   type DeserializedRecord,
   type SerializedRecord,
-} from '../../serialization'
+} from '../../../serialization'
 
-import { Color } from '../color'
-import { ControlDefinition, serialize, type SchemaType } from '../definition'
-import { type SendMessage } from '../instance'
+import { Color } from '../../color'
+import { ControlDefinition, serialize, type SchemaType } from '../../definition'
+import { type SendMessage } from '../../instance'
 
 import * as StyleSchema from './schema'
 import { StyleControl } from './style-control'
